Refactor identity fixtures
