+++ p2panda-rs/src/test_utils/fixtures/new_fixtures/operation_fixtures.rs
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
 use std::convert::TryFrom;
 
 use rstest::fixture;
